+++ src/mdio/core/v1/__init__.py
@@ -5,13 +5,15 @@ Exposes the MDIO overloads and core v1 functionality.
 
 from ._overloads import mdio
 from .builder import Builder
-from .constructor import write_mdio_metadata
+from .builder import write_mdio_metadata
+from ._serializer import (
+    make_coordinate,
+    make_dataset,
+    make_dataset_metadata,
+    make_named_dimension,
+    make_variable,
+)
 from .factory import AbstractTemplateFactory
-from .factory import make_coordinate
-from .factory import make_dataset
-from .factory import make_dataset_metadata
-from .factory import make_named_dimension
-from .factory import make_variable
 
 
 __all__ = [
