+++ tests/integration/test_v1_constructor.py
@@ -4,11 +4,11 @@ from datetime import datetime
 
 import numpy as np
 
-from mdio.core.v1.constructor import write_mdio_metadata
-from mdio.core.v1.factory import make_dataset
-from mdio.core.v1.factory import make_dataset_metadata
-from mdio.core.v1.factory import make_named_dimension
-from mdio.core.v1.factory import make_variable
+from mdio.core.v1.builder import write_mdio_metadata
+from mdio.core.v1._serializer import make_dataset
+from mdio.core.v1._serializer import make_dataset_metadata
+from mdio.core.v1._serializer import make_named_dimension
+from mdio.core.v1._serializer import make_variable
 from mdio.schema.compressors import ZFP
 from mdio.schema.compressors import Blosc
 from mdio.schema.dtype import ScalarType
