+++ tests/unit/schema/v1/test_template_builder.py
@@ -6,7 +6,7 @@ import pytest
 
 from mdio.core.v1.builder import Builder as TemplateBuilder
 from mdio.core.v1.builder import _BuilderState
-from mdio.core.v1.constructor import write_mdio_metadata
+from mdio.core.v1.builder import write_mdio_metadata
 from mdio.schema.compressors import Blosc
 from mdio.schema.dtype import ScalarType
 from mdio.schema.dtype import StructuredType
